requirements.txt
ctransformers==0.2.5
sentence-transformers==3.2.1
pinecone-client==4.1.2
langchain==0.2.17
langchain-pinecone==0.1.2
langchain-community==0.2.18
langchain-huggingface==0.0.3
streamlit==1.39.0
pypdf==4.3.1
python-dotenv==1.0.1
